adjust accelerator

+++ crates/nari2d_core/src/geometry/scale2d.rs
@@ -4,20 +4,20 @@ use std::ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Sub, SubAssign};
 
 #[derive(Copy, Clone, Default, Debug, PartialOrd, PartialEq)]
 pub struct Scale2d {
-    x: f64,
-    y: f64,
+    x: f32,
+    y: f32,
 }
 
 impl Scale2d {
     #[inline]
     #[must_use]
-    pub fn new(x: f64, y: f64) -> Self {
+    pub fn new(x: f32, y: f32) -> Self {
         Scale2d { x, y }
     }
 
     #[inline]
     #[must_use]
-    pub fn splat(v: f64) -> Self {
+    pub fn splat(v: f32) -> Self {
         Scale2d { x: v, y: v }
     }
 
@@ -29,41 +29,41 @@ impl Scale2d {
 
     #[inline]
     #[must_use]
-    pub fn to_array(self) -> [f64; 2] {
+    pub fn to_array(self) -> [f32; 2] {
         [self.x, self.y]
     }
 
     #[inline]
     #[must_use]
-    pub fn to_tuple(self) -> (f64, f64) {
+    pub fn to_tuple(self) -> (f32, f32) {
         (self.x, self.y)
     }
 
     #[inline]
     #[must_use]
-    pub fn to_vec(self) -> Vec<f64> {
+    pub fn to_vec(self) -> Vec<f32> {
         vec![self.x, self.y]
     }
 
     #[inline]
     #[must_use]
-    pub fn x(self) -> f64 {
+    pub fn x(self) -> f32 {
         self.x
     }
 
     #[inline]
-    pub fn set_x(&mut self, new_x: f64) {
+    pub fn set_x(&mut self, new_x: f32) {
         self.x = new_x;
     }
 
     #[inline]
     #[must_use]
-    pub fn y(self) -> f64 {
+    pub fn y(self) -> f32 {
         self.y
     }
 
     #[inline]
-    pub fn set_y(&mut self, new_y: f64) {
+    pub fn set_y(&mut self, new_y: f32) {
         self.y = new_y;
     }
 
@@ -130,8 +130,8 @@ impl Scale2d {
 
     #[inline]
     #[must_use]
-    pub fn linear_interpolate(self, end: Scale2d, along: f64) -> Self {
-        let along = along.clamp(0_f64, 1_f64);
+    pub fn linear_interpolate(self, end: Scale2d, along: f32) -> Self {
+        let along = along.clamp(0_f32, 1_f32);
         Scale2d {
             x: self.x + (end.x - self.x) * along,
             y: self.y + (end.y - self.y) * along,
@@ -194,10 +194,10 @@ impl Sub for Scale2d {
     }
 }
 
-impl Add<f64> for Scale2d {
+impl Add<f32> for Scale2d {
     type Output = Scale2d;
     #[inline]
-    fn add(self, rhs: f64) -> Self::Output {
+    fn add(self, rhs: f32) -> Self::Output {
         Scale2d {
             x: self.x + rhs,
             y: self.y + rhs,
@@ -205,11 +205,11 @@ impl Add<f64> for Scale2d {
     }
 }
 
-impl Sub<f64> for Scale2d {
+impl Sub<f32> for Scale2d {
     type Output = Scale2d;
 
     #[inline]
-    fn sub(self, rhs: f64) -> Self::Output {
+    fn sub(self, rhs: f32) -> Self::Output {
         Scale2d {
             x: self.x - rhs,
             y: self.y - rhs,
@@ -217,10 +217,10 @@ impl Sub<f64> for Scale2d {
     }
 }
 
-impl Mul<f64> for Scale2d {
+impl Mul<f32> for Scale2d {
     type Output = Scale2d;
     #[inline]
-    fn mul(self, rhs: f64) -> Self::Output {
+    fn mul(self, rhs: f32) -> Self::Output {
         Scale2d {
             x: self.x * rhs,
             y: self.y * rhs,
@@ -228,10 +228,10 @@ impl Mul<f64> for Scale2d {
     }
 }
 
-impl Div<f64> for Scale2d {
+impl Div<f32> for Scale2d {
     type Output = Scale2d;
     #[inline]
-    fn div(self, rhs: f64) -> Self::Output {
+    fn div(self, rhs: f32) -> Self::Output {
         Scale2d {
             x: self.x / rhs,
             y: self.y / rhs,
@@ -255,33 +255,33 @@ impl SubAssign for Scale2d {
     }
 }
 
-impl AddAssign<f64> for Scale2d {
+impl AddAssign<f32> for Scale2d {
     #[inline]
-    fn add_assign(&mut self, rhs: f64) {
+    fn add_assign(&mut self, rhs: f32) {
         self.set_x(self.x() + rhs);
         self.set_y(self.y() + rhs);
     }
 }
 
-impl SubAssign<f64> for Scale2d {
+impl SubAssign<f32> for Scale2d {
     #[inline]
-    fn sub_assign(&mut self, rhs: f64) {
+    fn sub_assign(&mut self, rhs: f32) {
         self.set_x(self.x() - rhs);
         self.set_y(self.y() - rhs);
     }
 }
 
-impl MulAssign<f64> for Scale2d {
+impl MulAssign<f32> for Scale2d {
     #[inline]
-    fn mul_assign(&mut self, rhs: f64) {
+    fn mul_assign(&mut self, rhs: f32) {
         self.set_x(self.x() * rhs);
         self.set_y(self.y() * rhs);
     }
 }
 
-impl DivAssign<f64> for Scale2d {
+impl DivAssign<f32> for Scale2d {
     #[inline]
-    fn div_assign(&mut self, rhs: f64) {
+    fn div_assign(&mut self, rhs: f32) {
         self.set_x(self.x() / rhs);
         self.set_y(self.y() / rhs);
     }
@@ -311,10 +311,10 @@ impl Sub for &Scale2d {
     }
 }
 
-impl Add<f64> for &Scale2d {
+impl Add<f32> for &Scale2d {
     type Output = Scale2d;
     #[inline]
-    fn add(self, rhs: f64) -> Self::Output {
+    fn add(self, rhs: f32) -> Self::Output {
         Scale2d {
             x: self.x + rhs,
             y: self.y + rhs,
@@ -322,10 +322,10 @@ impl Add<f64> for &Scale2d {
     }
 }
 
-impl Sub<f64> for &Scale2d {
+impl Sub<f32> for &Scale2d {
     type Output = Scale2d;
     #[inline]
-    fn sub(self, rhs: f64) -> Self::Output {
+    fn sub(self, rhs: f32) -> Self::Output {
         Scale2d {
             x: self.x - rhs,
             y: self.y - rhs,
@@ -333,10 +333,10 @@ impl Sub<f64> for &Scale2d {
     }
 }
 
-impl Mul<f64> for &Scale2d {
+impl Mul<f32> for &Scale2d {
     type Output = Scale2d;
     #[inline]
-    fn mul(self, rhs: f64) -> Self::Output {
+    fn mul(self, rhs: f32) -> Self::Output {
         Scale2d {
             x: self.x * rhs,
             y: self.y * rhs,
@@ -344,10 +344,10 @@ impl Mul<f64> for &Scale2d {
     }
 }
 
-impl Div<f64> for &Scale2d {
+impl Div<f32> for &Scale2d {
     type Output = Scale2d;
     #[inline]
-    fn div(self, rhs: f64) -> Self::Output {
+    fn div(self, rhs: f32) -> Self::Output {
         Scale2d {
             x: self.x / rhs,
             y: self.y / rhs,
@@ -355,14 +355,14 @@ impl Div<f64> for &Scale2d {
     }
 }
 
-impl From<[f64; 2]> for Scale2d {
-    fn from(from: [f64; 2]) -> Self {
+impl From<[f32; 2]> for Scale2d {
+    fn from(from: [f32; 2]) -> Self {
         Scale2d::new(from[0], from[1])
     }
 }
 
-impl From<(f64, f64)> for Scale2d {
-    fn from(from: (f64, f64)) -> Self {
+impl From<(f32, f32)> for Scale2d {
+    fn from(from: (f32, f32)) -> Self {
         Scale2d::new(from.0, from.1)
     }
 }
